requirements.txt
# Core dependencies
python-dotenv>=1.0.0           # 환경변수 관리
google-generativeai>=0.5.0     # Google Gemini API 연동
google-genai>=0.2.0            # Google Gemini Client API (멀티모달 분석용)

# Interface & Integration
discord.py>=2.3.2              # Discord 봇 인터페이스
notion-client>=2.2.0           # Notion API 연동

# Multimodal Analysis
Pillow>=10.0.0                 # 이미지 처리 (ImageAnalysisTool용)
python-docx>=1.0.0             # Word 문서 분석 (DocumentAnalysisTool용)
pdfplumber>=0.10.0             # PDF 분석 (PDFAnalysisTool용)

# AI & Machine Learning
torch>=2.1.0                   # 딥러닝 프레임워크 (임베딩 처리용)
transformers>=4.51.0           # Hugging Face 트랜스포머 모델
numpy>=1.24.0                  # 수치 연산 (벡터 처리용)
faiss-cpu>=1.8.0               # 벡터 유사도 검색

# Utilities
send2trash>=1.8.2              # 안전한 파일 삭제

# Multi-Agent Framework
crewai>=0.28.0                 # 멀티 에이전트 프레임워크
crewai-tools>=0.2.0            # CrewAI 도구 라이브러리

# Development & Testing
pytest>=7.4.0                  # 테스트 프레임워크
